refactor: marker is mdc
