+++ src/diagnostic/fastrace.rs
